Renamed folders and removed unused files
