+++ SeleniumTests/README.md
@@ -1,2 +1,4 @@
-## SampleWebApplication Web Project
-Test github webhook
+## Lattice Selenium Web Testing
+Open project in SeleniumTests folder in IDE such as IntelliJ IDEA.
+Locate the FrontendTests.java in SeleniumTests/src/test/java/com/seleniumtests/journey/FrontendTests.java
+From here you will be able to run all the tests.
