+++ SeleniumTests/src/main/java/com/pagesandframework/pages/Page.java
@@ -1,4 +1,4 @@
-package com.sampleweb.pages;
+package com.pagesandframework.pages;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.StaleElementReferenceException;
