+++ SeleniumTests/src/test/java/com/seleniumtests/listeners/TestListener.java
@@ -2,7 +2,7 @@
 Placeholder for the listener
  */
 
-package com.sampleweb.listeners;
+package com.seleniumtests.listeners;
 
 import org.testng.ITestListener;
 import org.testng.TestListenerAdapter;
