+++ SeleniumTests/src/test/resources/test.properties
@@ -1,3 +1,2 @@
 test.local=true
-test.gridUrl=http://34.254.183.221:4444/wd/hub
 test.mainUrl=http://localhost:4200/
