--- simple-webapp-sample/README.md
@@ -1,2 +1,0 @@
-## SampleWebApplication Web Project
-Test github webhook
