--- simple-webapp-sample/src/main/java/com/sampleweb/pages/GiftCardPage.java
@@ -1,17 +1,0 @@
-package com.sampleweb.pages;
-
-import org.openqa.selenium.By;
-import org.openqa.selenium.remote.RemoteWebDriver;
-
-public class GiftCardPage extends Page{
-
-    private String pageTitle = "";
-
-    public GiftCardPage(RemoteWebDriver driver) {
-        super(driver);
-    }
-
-    public boolean isLoaded() {
-        return driver.findElement(By.xpath(pageTitle)).isDisplayed();
-    }
-}
