--- simple-webapp-sample/src/main/resources/testng.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<suite name="GRID SAMPLE TEST" parallel="tests" thread-count="3">
-    <test name="GRID TEST WITH SERIAL EXECUTION WITH BROWSER FF">
-        <parameter name ="browserType" value="firefox"/>
-        <classes>
-            <class name ="com.sampleweb.journey.GridExample"/>
-        </classes>
-    </test>
-    <!-- <test name="GRID TEST WITH SERIAL EXECTION WITH BROWSER CHROME">
-        <parameter name ="browserType" value="chrome"/>
-        <classes>
-            <class name ="com.sampleweb.journey.GridExample"/>
-        </classes>
-    </test> -->
-</suite>
