--- simple-webapp-sample/src/main/webapp/index.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>Hello</title>
-</head>
-<body>
-    <h2>Hello World! Wednesday Try!</h2>
-</body>
-</html>
