--- simple-webapp-sample/src/test/java/com/sampleweb/listeners/TestListener.java
@@ -1,12 +1,0 @@
-/*
-Placeholder for the listener
- */
-
-package com.sampleweb.listeners;
-
-import org.testng.ITestListener;
-import org.testng.TestListenerAdapter;
-
-public class TestListener extends TestListenerAdapter implements ITestListener {
-
-}
