--- simple-webapp-sample/src/test/resources/allure.properties
@@ -1,2 +1,0 @@
-allure.results.directory=allure-results
-allure.report.directory=${basedir}/target/site/allure-maven-plugin
